[UPDATE,REFACTOR] 파일구조 변경 및 Dialog 추가
- PreInfoAddDialog, PreInfoDeleteDialog 추가
- RegisterPopover 파일만 추가
- signinPopoverPage 추가
- 헤더에 전역 color 추가
- 관리메뉴와 시간표메뉴 한 파일에 묶음
- 시간표목록페이지, 메뉴목록페이지 파일만 추가
