--- src/components/templates/AuthManagement/AuthListView.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import ContentHeader from '../../modules/ContentHeader/ContentHeader';
-
-const Wrapper = styled.div``;
-
-const AuthListView = () => {
-  return (
-    <Wrapper>
-      <ContentHeader title="권한 관리" />
-    </Wrapper>
-  );
-};
-
-export default AuthListView;
