+++ src/components/templates/Header/Auth/RegisterPopover.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const RegisterPage = () => {
+  return <></>;
+};
+
+export default RegisterPage;
