+++ src/components/templates/Header/Header.js
@@ -3,25 +3,18 @@ import PropTypes from 'prop-types';
 import AppBar from '../../modules/AppBar/AppBar';
 import SideMenu from '../../modules/SideMenu/SideMenu';
 
-const Header = ({ open, sideMenuOpen, sideMenuClose, menuItems, path }) => {
+const Header = ({ open, setOpen, menuItems }) => {
   return (
     <>
-      <AppBar
-        open={open}
-        sideMenuOpen={sideMenuOpen}
-        sideMenuClose={sideMenuClose}
-        path={path}
-      />
-      <SideMenu open={open} items={menuItems} path={path} />
+      <AppBar open={open} setOpen={setOpen} />
+      <SideMenu open={open} items={menuItems} />
     </>
   );
 };
 Header.propTypes = {
   open: PropTypes.bool.isRequired,
-  sideMenuOpen: PropTypes.func.isRequired,
-  sideMenuClose: PropTypes.func.isRequired,
-  menuItems: PropTypes.arrayOf(PropTypes.array).isRequired,
-  path: PropTypes.string.isRequired
+  setOpen: PropTypes.func.isRequired,
+  menuItems: PropTypes.arrayOf(PropTypes.array).isRequired
 };
 
 export default Header;
