--- src/components/templates/LogManagement/LogListView.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import ContentHeader from '../../modules/ContentHeader/ContentHeader';
-
-const Wrapper = styled.div``;
-
-const LogListView = () => {
-  return (
-    <Wrapper>
-      <ContentHeader title="로그 관리" />
-    </Wrapper>
-  );
-};
-
-export default LogListView;
