+++ src/components/templates/Manage/AuthManagement/AuthListPage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+import ContentHeader from '../../../modules/ContentHeader/ContentHeader';
+
+const Wrapper = styled.div``;
+
+const AuthListView = () => {
+  return (
+    <Wrapper>
+      <ContentHeader title="권한 관리" />
+    </Wrapper>
+  );
+};
+
+export default AuthListView;
