+++ src/components/templates/Manage/JobInfoUpload/JobInfoUploadPage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+import ContentHeader from '../../../modules/ContentHeader/ContentHeader';
+
+const Wrapper = styled.div``;
+
+const JobInfoUploadView = () => {
+  return (
+    <Wrapper>
+      <ContentHeader title="취업정보 업로드" />
+    </Wrapper>
+  );
+};
+
+export default JobInfoUploadView;
