+++ src/components/templates/Manage/NoticeManagement/NoticeListPage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+import ContentHeader from '../../../modules/ContentHeader/ContentHeader';
+
+const Wrapper = styled.div``;
+
+const NoticeListView = () => {
+  return (
+    <Wrapper>
+      <ContentHeader title="알림 관리" />
+    </Wrapper>
+  );
+};
+
+export default NoticeListView;
