+++ src/components/templates/Manage/PostManagement/PostMovePage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+import ContentHeader from '../../../modules/ContentHeader/ContentHeader';
+
+const Wrapper = styled.div``;
+
+const PostMoveView = () => {
+  return (
+    <Wrapper>
+      <ContentHeader title="게시글 이동" />
+    </Wrapper>
+  );
+};
+
+export default PostMoveView;
