+++ src/components/templates/Manage/ReportManagement/ReportListPage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+import ContentHeader from '../../../modules/ContentHeader/ContentHeader';
+
+const Wrapper = styled.div``;
+
+const ReportListView = () => {
+  return (
+    <Wrapper>
+      <ContentHeader title="신고 관리" />
+    </Wrapper>
+  );
+};
+
+export default ReportListView;
