+++ src/components/templates/Manage/StatisticsManagement/StatisticsPage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+import ContentHeader from '../../../modules/ContentHeader/ContentHeader';
+
+const Wrapper = styled.div``;
+
+const StatisticsView = () => {
+  return (
+    <Wrapper>
+      <ContentHeader title="통계 관리" />
+    </Wrapper>
+  );
+};
+
+export default StatisticsView;
