+++ src/components/templates/Manage/TagManagement/TagListPage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+import ContentHeader from '../../../modules/ContentHeader/ContentHeader';
+
+const Wrapper = styled.div``;
+
+const TagListView = () => {
+  return (
+    <Wrapper>
+      <ContentHeader title="태그 관리" />
+    </Wrapper>
+  );
+};
+
+export default TagListView;
