--- src/components/templates/ReportManagement/ReportListView.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import ContentHeader from '../../modules/ContentHeader/ContentHeader';
-
-const Wrapper = styled.div``;
-
-const ReportListView = () => {
-  return (
-    <Wrapper>
-      <ContentHeader title="신고 관리" />
-    </Wrapper>
-  );
-};
-
-export default ReportListView;
