--- src/components/templates/StatisticsManagement/StatisticsView.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import ContentHeader from '../../modules/ContentHeader/ContentHeader';
-
-const Wrapper = styled.div``;
-
-const StatisticsView = () => {
-  return (
-    <Wrapper>
-      <ContentHeader title="통계 관리" />
-    </Wrapper>
-  );
-};
-
-export default StatisticsView;
