--- src/components/templates/TagManagement/TagListView.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import ContentHeader from '../../modules/ContentHeader/ContentHeader';
-
-const Wrapper = styled.div``;
-
-const TagListView = () => {
-  return (
-    <Wrapper>
-      <ContentHeader title="태그 관리" />
-    </Wrapper>
-  );
-};
-
-export default TagListView;
